refactor: move packages from usecase to usecase.customer and usecase.employee
