+++ src/main/java/com/postech/fastfood/config/UseCaseBeanConfiguration.java
@@ -6,7 +6,11 @@ import com.postech.fastfood.core.service.customer.CreateCustomerWithNameAndEmail
 import com.postech.fastfood.core.service.customer.FindCustomerByCpfUseCaseImpl;
 import com.postech.fastfood.core.service.customer.FindCustomerrByEmailUseCaseImpl;
 import com.postech.fastfood.core.service.employee.CreateEmployeeUseCaseImpl;
-import com.postech.fastfood.core.usecase.*;
+import com.postech.fastfood.core.usecase.FindUserByCpfUseCase;
+import com.postech.fastfood.core.usecase.FindUserByEmailUseCase;
+import com.postech.fastfood.core.usecase.customer.CreateCustomerWithCpfUseCase;
+import com.postech.fastfood.core.usecase.customer.CreateCustomerWithNameAndEmailUseCase;
+import com.postech.fastfood.core.usecase.employee.CreateEmployeeUseCase;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -14,27 +18,27 @@ import org.springframework.context.annotation.Configuration;
 public class UseCaseBeanConfiguration {
 
     @Bean
-    public CreateCustomerWithNameAndEmailUseCase createCustomerWithNameAndEmailUseCase(UserRepositoryPort userRepositoryPort){
+    public CreateCustomerWithNameAndEmailUseCase createCustomerWithNameAndEmailUseCase(UserRepositoryPort userRepositoryPort) {
         return new CreateCustomerWithNameAndEmailUseCaseImpl(userRepositoryPort);
     }
 
     @Bean
-    public CreateCustomerWithCpfUseCase createCustomerWithCpfUseCase(UserRepositoryPort userRepositoryPort){
+    public CreateCustomerWithCpfUseCase createCustomerWithCpfUseCase(UserRepositoryPort userRepositoryPort) {
         return new CreateCustomerWithCpfUseCaseImpl(userRepositoryPort);
     }
 
     @Bean
-    public FindUserByEmailUseCase findCustomerByEmailUseCase(UserRepositoryPort userRepositoryPort){
+    public FindUserByEmailUseCase findCustomerByEmailUseCase(UserRepositoryPort userRepositoryPort) {
         return new FindCustomerrByEmailUseCaseImpl(userRepositoryPort);
     }
 
     @Bean
-    public FindUserByCpfUseCase findCustomerByCpfUseCase (UserRepositoryPort userRepositoryPort){
+    public FindUserByCpfUseCase findCustomerByCpfUseCase(UserRepositoryPort userRepositoryPort) {
         return new FindCustomerByCpfUseCaseImpl(userRepositoryPort);
     }
 
     @Bean
-    public CreateEmployeeUseCase createEmployeeUseCase(UserRepositoryPort userRepositoryPort){
+    public CreateEmployeeUseCase createEmployeeUseCase(UserRepositoryPort userRepositoryPort) {
         return new CreateEmployeeUseCaseImpl(userRepositoryPort);
     }
