+++ src/main/java/com/postech/fastfood/core/service/employee/CreateEmployeeUseCaseImpl.java
@@ -4,7 +4,7 @@ import com.postech.fastfood.core.domain.Employee;
 import com.postech.fastfood.core.exception.CpfAlreadyInUseException;
 import com.postech.fastfood.core.exception.EmailAlreadyExistsException;
 import com.postech.fastfood.core.ports.UserRepositoryPort;
-import com.postech.fastfood.core.usecase.CreateEmployeeUseCase;
+import com.postech.fastfood.core.usecase.employee.CreateEmployeeUseCase;
 import org.springframework.dao.DataIntegrityViolationException;
 
 public class CreateEmployeeUseCaseImpl implements CreateEmployeeUseCase {
@@ -17,10 +17,10 @@ public class CreateEmployeeUseCaseImpl implements CreateEmployeeUseCase {
 
     @Override
     public Employee execute(Employee user) {
-        Employee userSaved= null;
+        Employee userSaved = null;
         try {
-            user.setCpf(user.getCpf().replace(".",""));
-            user.setCpf(user.getCpf().replace("-",""));
+            user.setCpf(user.getCpf().replace(".", ""));
+            user.setCpf(user.getCpf().replace("-", ""));
             userSaved = (Employee) this.userRepositoryPort.save(user);
         } catch (DataIntegrityViolationException e) {
             String message = e.getMostSpecificCause().getMessage();
