+++ src/main/java/com/postech/fastfood/core/usecase/employee/CreateEmployeeUseCase.java
@@ -1,4 +1,4 @@
-package com.postech.fastfood.core.usecase;
+package com.postech.fastfood.core.usecase.employee;
 
 import com.postech.fastfood.core.domain.Employee;
 
@@ -6,4 +6,4 @@ public interface CreateEmployeeUseCase {
 
     Employee execute(Employee employee);
 
-    }
+}
